+++ R/eyelinker.R
@@ -5,21 +5,21 @@
 #' easy as possible.
 #'
 #' For documentation of the structure of the returned data, see the "format" vignette:
+#' * `vignette("format", package = "eyelinker")`
 #'
-#' \code{vignette("format", package = "eyelinker")}
-#'
-#' For worked examples illustrating the package in action, see the "basics" vignette:
-#'
-#' \code{vignette("basics", package = "eyelinker")}
-#' @docType package
-#' @name eyelinker
-NULL
+#' For worked examples illustrating the package in action, see the "basics",
+#' "epoching", and "batch_import" vignettes:
+#' * `vignette("basics", package = "eyelinker")`
+#' * `vignette("epoching", package = "eyelinker")`
+#' * `vignette("batch_import", package = "eyelinker")`
+#' @md
+"_PACKAGE"
 
 #' @importFrom stats median
 #' @importFrom utils tail
 #' @importFrom stringr str_sub<- str_detect fixed
 #' @importFrom tibble as_tibble add_column
-#' @importFrom readr read_lines read_tsv read_table2
+#' @importFrom readr read_lines read_tsv read_table
 #' @importFrom stringi stri_enc_toascii stri_split_regex stri_split_fixed stri_count_fixed
 #' @importFrom intervals which_nearest distance_to_nearest Intervals
 NULL