+++ NEWS.md
@@ -1,3 +1,11 @@
+# eyelinker 0.2.2
+
+ * Set tidyverse-style aliases to be preferred style in manual and vignettes
+ * Added new vignette on epoching data based on message events
+ * Added new batch import vignette
+ * Modernized syntax in 'Getting Started' vignette
+ * Fixed compatibility with latest readr (#5, by Jenny Bryan)
+
 # eyelinker 0.2.1
 
  * Fixed bug reading screen resolution from certain .asc files